reorganize enums
