+++ testgres/__init__.py
@@ -3,13 +3,13 @@ from .backup import NodeBackup
 from .config import TestgresConfig, configure_testgres
 
 from .connection import \
-    IsolationLevel, \
     NodeConnection, \
     InternalError, \
     ProgrammingError
 
 from .exceptions import *
-from .node import NodeStatus, PostgresNode
+from .enums import *
+from .node import PostgresNode
 
 from .utils import \
     reserve_port, \
